+++ tests/files/no-unused-modules/flow-1.js
@@ -1,2 +1,0 @@
-// @flow strict
-export type Bar = number;