+++ tests/files/no-unused-modules/flow-2.js
@@ -1,2 +1,0 @@
-// @flow strict
-export type FooType = string;