+++ tests/src/rules/no-restricted-paths.js
@@ -173,10 +173,11 @@ ruleTester.run('no-restricted-paths', rule, {
           target: './tests/files/restricted-paths/server/one',
           from: './tests/files/restricted-paths/server',
           except: ['./one'],
+          message: 'Custom message',
         } ],
       } ],
       errors: [ {
-        message: 'Unexpected path "../two/a.js" imported in restricted zone.',
+        message: 'Unexpected path "../two/a.js" imported in restricted zone. Custom message',
         line: 1,
         column: 15,
       } ],