+++ tests/src/rules/no-unused-modules.js
@@ -1,4 +1,4 @@
-import { test, testFilePath } from '../utils'
+import { test, testFilePath, getTSParsers } from '../utils'
 import jsxConfig from '../../../config/react'
 import typescriptConfig from '../../../config/typescript'
 
@@ -442,6 +442,31 @@ ruleTester.run('no-unused-modules', rule, {
   invalid: [],
 })
 
+describe('renameDefault', () => {
+  ruleTester.run('no-unused-modules', rule, {
+    valid: [
+      test({ options: unusedExportsOptions,
+        code: 'export { default as Component } from "./Component"',
+        filename: testFilePath('./no-unused-modules/renameDefault/components.js')}),
+      test({ options: unusedExportsOptions,
+        code: 'export default function Component() {}',
+        filename: testFilePath('./no-unused-modules/renameDefault/Component.js')}),
+    ],
+    invalid: [],
+  })
+  ruleTester.run('no-unused-modules', rule, {
+    valid: [
+      test({ options: unusedExportsOptions,
+        code: 'export { default as ComponentA } from "./ComponentA";export { default as ComponentB } from "./ComponentB";',
+        filename: testFilePath('./no-unused-modules/renameDefault-2/components.js')}),
+      test({ options: unusedExportsOptions,
+        code: 'export default function ComponentA() {};',
+        filename: testFilePath('./no-unused-modules/renameDefault-2/ComponentA.js')}),
+    ],
+    invalid: [],
+  })
+})
+
 describe('test behaviour for new file', () => {
   before(() => {
     fs.writeFileSync(testFilePath('./no-unused-modules/file-added-0.js'), '', {encoding: 'utf8'})
@@ -680,38 +705,6 @@ describe('do not report unused export for files mentioned in package.json', () =
   })
 })
 
-describe('correctly report flow types', () => {
-  ruleTester.run('no-unused-modules', rule, {
-    valid: [
-      test({
-        options: unusedExportsOptions,
-        code: 'import { type FooType } from "./flow-2";',
-        parser: require.resolve('babel-eslint'),
-        filename: testFilePath('./no-unused-modules/flow-0.js'),
-      }),
-      test({
-        options: unusedExportsOptions,
-        code: `// @flow strict
-               export type FooType = string;`,
-        parser: require.resolve('babel-eslint'),
-        filename: testFilePath('./no-unused-modules/flow-2.js'),
-      }),
-    ],
-    invalid: [
-      test({
-        options: unusedExportsOptions,
-        code: `// @flow strict
-               export type Bar = string;`,
-        parser: require.resolve('babel-eslint'),
-        filename: testFilePath('./no-unused-modules/flow-1.js'),
-        errors: [
-          error(`exported declaration 'Bar' not used within other modules`),
-        ],
-      }),
-    ],
-  })
-})
-
 describe('Avoid errors if re-export all from umd compiled library', () => {
   ruleTester.run('no-unused-modules', rule, {
     valid: [
@@ -743,7 +736,78 @@ describe('correctly work with Typescript only files', () => {
           error(`exported declaration 'b' not used within other modules`),
         ],
       }),
+      test({
+        options: unusedExportsTypescriptOptions,
+        code: `export interface c {};`,
+        parser: require.resolve('babel-eslint'),
+        filename: testFilePath('./no-unused-modules/typescript/file-ts-c.ts'),
+        errors: [
+          error(`exported declaration 'c' not used within other modules`),
+        ],
+      }),
+      test({
+        options: unusedExportsTypescriptOptions,
+        code: `export type d = {};`,
+        parser: require.resolve('babel-eslint'),
+        filename: testFilePath('./no-unused-modules/typescript/file-ts-d.ts'),
+        errors: [
+          error(`exported declaration 'd' not used within other modules`),
+        ],
+      }),
     ],
+  })
+})
+
+context('TypeScript', function () {
+  getTSParsers().forEach((parser) => {
+    typescriptRuleTester.run('no-unused-modules', rule, {
+      valid: [
+        test({
+          options: unusedExportsTypescriptOptions,
+          code: 'import a from "file-ts-a";',
+          parser: parser,
+          filename: testFilePath('./no-unused-modules/typescript/file-ts-a.ts'),
+        }),
+      ],
+      invalid: [
+        test({
+          options: unusedExportsTypescriptOptions,
+          code: `export const b = 2;`,
+          parser: parser,
+          filename: testFilePath('./no-unused-modules/typescript/file-ts-b.ts'),
+          errors: [
+            error(`exported declaration 'b' not used within other modules`),
+          ],
+        }),
+        test({
+          options: unusedExportsTypescriptOptions,
+          code: `export interface c {};`,
+          parser: parser,
+          filename: testFilePath('./no-unused-modules/typescript/file-ts-c.ts'),
+          errors: [
+            error(`exported declaration 'c' not used within other modules`),
+          ],
+        }),
+        test({
+          options: unusedExportsTypescriptOptions,
+          code: `export type d = {};`,
+          parser: parser,
+          filename: testFilePath('./no-unused-modules/typescript/file-ts-d.ts'),
+          errors: [
+            error(`exported declaration 'd' not used within other modules`),
+          ],
+        }),
+        test({
+          options: unusedExportsTypescriptOptions,
+          code: `export enum e { f };`,
+          parser: parser,
+          filename: testFilePath('./no-unused-modules/typescript/file-ts-e.ts'),
+          errors: [
+            error(`exported declaration 'e' not used within other modules`),
+          ],
+        }),
+      ],
+    })
   })
 })